Move graph-degradation into a subdirectory (review feedback, #166)
Per review on #195: group the degradation subsystem under js/ui/graph-degradation/
the way js/ui/overlays/ is organized, with index.js as the wiring/public entry.

  js/ui/graph-degradation/
    index.js            (was graph-degradation.js — wiring + public API)
    health-plasma.js
    deck-perturbation.js
    params.js           (was graph-degradation-params.js)
    params.test.js
    index.test.js       (was graph-degradation.test.js)

Pure move + import-path updates: internal sibling imports stay ./, external deps
go ../, and the two consumers (visual-renderer, preview) import
./graph-degradation/index.js. Public API (initGraphDegradation / updateFromState /
stopGraphDegradation) unchanged. make check green (1105 tests).
