+++ js/ui/graph-degradation/health-plasma.js
