+++ js/ui/graph-degradation/params.js
